# perf: vreplication client CPU usage