--- a/go/mysql/query.go
+++ b/go/mysql/query.go
@@ -253,12 +253,15 @@ func (c *Conn) readColumnDefinitionType(field *querypb.Field, index int) error {
 
 // parseRow parses an individual row.
 // Returns a SQLError.
-func (c *Conn) parseRow(data []byte, fields []*querypb.Field, reader func([]byte, int) ([]byte, int, bool)) ([]sqltypes.Value, error) {
+func (c *Conn) parseRow(data []byte, fields []*querypb.Field, reader func([]byte, int) ([]byte, int, bool), result []sqltypes.Value) ([]sqltypes.Value, error) {
 	colNumber := len(fields)
-	result := make([]sqltypes.Value, colNumber)
+	if result == nil {
+		result = make([]sqltypes.Value, 0, colNumber)
+	}
 	pos := 0
 	for i := 0; i < colNumber; i++ {
 		if data[pos] == NullValue {
+			result = append(result, sqltypes.Value{})
 			pos++
 			continue
 		}
@@ -268,7 +271,7 @@ func (c *Conn) parseRow(data []byte, fields []*querypb.Field, reader func([]byte
 		if !ok {
 			return nil, NewSQLError(CRMalformedPacket, SSUnknownSQLState, "decoding string failed")
 		}
-		result[i] = sqltypes.MakeTrusted(fields[i].Type, s)
+		result = append(result, sqltypes.MakeTrusted(fields[i].Type, s))
 	}
 	return result, nil
 }
@@ -464,7 +467,7 @@ func (c *Conn) ReadQueryResult(maxrows int, wantfields bool) (*sqltypes.Result, 
 		}
 
 		// Regular row.
-		row, err := c.parseRow(data, result.Fields, readLenEncStringAsBytesCopy)
+		row, err := c.parseRow(data, result.Fields, readLenEncStringAsBytesCopy, nil)
 		if err != nil {
 			c.recycleReadPacket()
 			return nil, false, 0, err