--- a/go/test/endtoend/messaging/message_test.go
+++ b/go/test/endtoend/messaging/message_test.go
@@ -104,7 +104,7 @@ func TestMessage(t *testing.T) {
 
 	// Consume first message.
 	start := time.Now().UnixNano()
-	got, err := streamConn.FetchNext()
+	got, err := streamConn.FetchNext(nil)
 	require.NoError(t, err)
 
 	want := []sqltypes.Value{
@@ -131,7 +131,7 @@ func TestMessage(t *testing.T) {
 	}
 
 	// Consume the resend.
-	_, err = streamConn.FetchNext()
+	_, err = streamConn.FetchNext(nil)
 	require.NoError(t, err)
 	qr = exec(t, conn, "select time_next, epoch from vitess_message where id = 1")
 	next, epoch = getTimeEpoch(qr)
@@ -219,7 +219,7 @@ func TestThreeColMessage(t *testing.T) {
 
 	exec(t, conn, "insert into vitess_message3(id, msg1, msg2) values(1, 'hello world', 3)")
 
-	got, err := streamConn.FetchNext()
+	got, err := streamConn.FetchNext(nil)
 	require.NoError(t, err)
 	want := []sqltypes.Value{
 		sqltypes.NewInt64(1),