--- a/go/vt/vttablet/tabletserver/throttle/client.go
+++ b/go/vt/vttablet/tabletserver/throttle/client.go
@@ -19,12 +19,29 @@ package throttle
 import (
 	"context"
 	"net/http"
+	"sync"
+	"sync/atomic"
 	"time"
 )
 
 const (
 	throttleCheckDuration = 250 * time.Millisecond
 )
+
+var throttleTicks int64
+var throttleInit sync.Once
+
+func initThrottleTicker() {
+	throttleInit.Do(func() {
+		go func() {
+			tick := time.NewTicker(throttleCheckDuration)
+			defer tick.Stop()
+			for range tick.C {
+				atomic.AddInt64(&throttleTicks, 1)
+			}
+		}()
+	})
+}
 
 // Client construct is used by apps who wish to consult with a throttler. It encapsulates the check/throttling/backoff logic
 type Client struct {
@@ -33,11 +50,12 @@ type Client struct {
 	checkType ThrottleCheckType
 	flags     CheckFlags
 
-	lastSuccessfulThrottleCheck time.Time
+	lastSuccessfulThrottle int64
 }
 
 // NewProductionClient creates a client suitable for foreground/production jobs, which have normal priority.
 func NewProductionClient(throttler *Throttler, appName string, checkType ThrottleCheckType) *Client {
+	initThrottleTicker()
 	return &Client{
 		throttler: throttler,
 		appName:   appName,
@@ -51,6 +69,7 @@ func NewProductionClient(throttler *Throttler, appName string, checkType Throttl
 // NewBackgroundClient creates a client suitable for background jobs, which have low priority over productio ntraffic,
 // e.g. migration, table pruning, vreplication
 func NewBackgroundClient(throttler *Throttler, appName string, checkType ThrottleCheckType) *Client {
+	initThrottleTicker()
 	return &Client{
 		throttler: throttler,
 		appName:   appName,
@@ -75,7 +94,7 @@ func (c *Client) ThrottleCheckOK(ctx context.Context) (throttleCheckOK bool) {
 		// no throttler
 		return true
 	}
-	if time.Since(c.lastSuccessfulThrottleCheck) <= throttleCheckDuration {
+	if c.lastSuccessfulThrottle >= atomic.LoadInt64(&throttleTicks) {
 		// if last check was OK just very recently there is no need to check again
 		return true
 	}
@@ -84,7 +103,7 @@ func (c *Client) ThrottleCheckOK(ctx context.Context) (throttleCheckOK bool) {
 	if checkResult.StatusCode != http.StatusOK {
 		return false
 	}
-	c.lastSuccessfulThrottleCheck = time.Now()
+	c.lastSuccessfulThrottle = atomic.LoadInt64(&throttleTicks)
 	return true
 
 }
@@ -105,10 +124,8 @@ func (c *Client) ThrottleCheckOKOrWait(ctx context.Context) bool {
 // The function is not thread safe.
 func (c *Client) Throttle(ctx context.Context) {
 	for {
-		select {
-		case <-ctx.Done():
+		if ctx.Err() != nil {
 			return
-		default:
 		}
 		if c.ThrottleCheckOKOrWait(ctx) {
 			break