## Description

Hi all! This week we're doing performance for the `vreplication` _client_. As discussed on our previous PR, the replication process is already bottlenecked on MySQL insert performance, so any optimizations we perform on the client at this point are not going to show up on wall time, but they're still very relevant because the replicator runs inside of the `vttablet`s, so reducing CPU usage here allows us to serve more queries concurrently with the replication process.

For this round of optimizations, I've simply chipped away at all the warm points in the profile for a `vttablet` client performing replication. Let's break down the individual optimizations in this PR, starting with the whole profile:

![szoter_annotated_image](https://user-images.githubusercontent.com/42793/116101029-c8011800-a6ad-11eb-8154-fea00e99f98a.jpeg)

I've annotated the warm spots in the profile in green. Let's go through them from left to right.

### 1. `mysql.(*Conn).readOnePacket`

There is an interesting "glitch" in this part of the flame graph that caught my attention. You can see that most of the calls into the runtime's allocation code are warranted (we are allocating a byte slice for each packet because the slice is later re-used as backing storage for the parsed rows -- this is efficient enough), but there's a significant amount of CPU time spent allocating memory in the middle column, under `mysql.(*Conn).readHeaderFrom`.

Looking at the code for this function, it's clear that it was designed to **not** allocate memory, with a tiny buffer on the stack, and yet it's a constant source of allocations:

https://github.com/vitessio/vitess/blob/c5cd3067aeb9d952a2f45084c37634267e4f9062/go/mysql/conn.go#L326-L355

The issue can be found trivially with escape analysis (running the compiler with `go build -gcflags="-m"`): the 4-byte array that is supposed to be in the stack is actually being allocated on the heap for each function call. This is because the Go compiler really struggles to perform escape analysis across interface calls (such as `io.Reader`). It can't tell that the subslice into the array is not being kept by the called code, so it needs to place it into the heap.

Unfortunately, we cannot monomorphise the `io.Reader` argument to help the compiler with escape analysis (because the `Conn` can be either a buffered or unbuffered conn), so we resort to the second best option: if we move the small array into `Conn` itself, it'll only be allocated once _per `Conn`_ as opposed to once per packet. https://github.com/vitessio/vitess/commit/e0b90daebe9e6b98d969934a24899b41d25e3a68

| Before | After |
|----------|--------|
| ![image](https://user-images.githubusercontent.com/42793/116101544-334aea00-a6ae-11eb-9bd6-dfb279a4fb80.png) |  ![image](https://user-images.githubusercontent.com/42793/116101870-7f962a00-a6ae-11eb-96e3-02feefb0b4c9.png) |

### 2. `mysql.(*Conn).parseRow`

The `parseRow` function in Vitess' `Conn` was already optimized very significantly in a previous PR: [**we stopped copying the contents from the row's backing storage when converting them into individual `Value`s**](https://github.com/vitessio/vitess/pull/7800/commits/91d9627d1f4dbe152f9a97e7774c86e68804ef94). However, we can still see in the flame graph that there is a significant amount of CPU usage in runtime allocation code. What is being allocated here is the actual slice that contains the `Value`s, i.e. `result []Value`, and this is a particularly wasteful allocation because the slice is always the exact same size for a query, so it could be reused over and over again while reading rows. 

Commit https://github.com/vitessio/vitess/commit/d6674b66779f3e02c32b045218b055927d0868e9 does just that, but taking special care towards backwards compatibility in the API. The `parseRow` function now takes an _optional_ target slice where the `Value`s will be stored, but still handles gracefully the case where the target slice is `nil` and allocates it to the required size. This allows users of the API to continue using the allocating version by passing `nil` instead of an existing slice. It also allows ergonomic usage by keeping the argument and return value of the function the same:

https://github.com/vitessio/vitess/blob/d6674b66779f3e02c32b045218b055927d0868e9/go/vt/vttablet/tabletserver/vstreamer/rowstreamer.go#L261-L264

With this calling design, the first time `FetchNext` is called, its argument will be `nil`, so a new slice will be allocated to store the rows, but in subsequent calls, the slice will be automatically reused. 

| Before | After |
|----------|--------|
| ![image](https://user-images.githubusercontent.com/42793/116105083-457a5780-a6b1-11eb-9636-7b0ce738e31b.png) | ![image](https://user-images.githubusercontent.com/42793/116105142-53c87380-a6b1-11eb-92c1-57a03e4ccd00.png) |

### 3. `sqltypes.RowToProto3`

The conversion of Vitess row data (`[]Value`) into the ProtoBuf form that will be sent over the wire is the most expensive part of the whole row streaming process. This is bad, particularly because there is actually no conversion to be performed here: it's just copying the raw data of the rows into the equivalent ProtoBuf-specified struct. As seen on the flame graph, the massive cost comes from allocating the same `Row` objects over and over again.

There are many ways to work around this performance issue. The most obvious one would be using an object pool to reuse `Row` objects between calls. This would be simple to implement but it has a shortcoming: a shared pool of `Row` objects would contain very different sizes of pooled objects, because the average size of an individual row varies wildly depending on the specific _table_ that is being currently copied. We can do better than a global pool if we scope down memory reuse to our local copying operation, where all the `Row` objects will be similarly sized, as they contain the exact same amount of columns.

Commit https://github.com/vitessio/vitess/commit/dafc1cb729d7be7dff2c05bd05a926005eb9a044 performs the local memory reuse without using a memory pool: instead, it uses a slice of `Row` objects, which is faster than a pool, also bounded (because there is a maximum size for the packets that the streamer will send) and it can be reused directly into the `Response` object. A new `sqltypes.RowToProto3Inplace` has been implemented to allow conversion without allocating a new `Row` object, and to minimize duplication, the old `RowToProto3` has been implemented on top of the new API.

| Before | After |
|----------|--------|
| ![image](https://user-images.githubusercontent.com/42793/116106882-ca19a580-a6b2-11eb-85aa-d7ce5009ae71.png) | ![image](https://user-images.githubusercontent.com/42793/116106927-d7369480-a6b2-11eb-8f60-c111b2a3a016.png) |

### 4. `throttle.(*Client).ThrottleCheckOK`

Ooh, here's a spicy one. It's not often that one sees a significant amount of CPU time spent in wall-clock time acquisition (i.e. `runtime.nanotime` for Go programs). Getting the system's time has been an extremely fast operation in Linux systems for many years; back in the day, clock access was gated behind a syscall, with the consequent overhead of context switching into kernel space. This is not the case anymore: time access for userland process is now available via a [vDSO](https://lwn.net/Articles/446528/), wherein the kernel maps the relevant clock data directly into our process' memory space so there is essentially no overhead when accessing it from our programs. Of course, the Go runtime takes advantage of this optimization (manually -- the Go runtime has to re-implement a lot of the vDSO logic that exists in `libc` because the Go runtime doesn't use `libc` :upside_down_face:), and yet we are spending _a lot_ of CPU time in our Throttler implementation just measuring the current time.

Why is this? Well, the Throttler is calling `time.Since` to make sure that we don't call the external throttler service too often. The throttler is effectively throttled at 250ms per call. How ironic. But it turns out that we do a lot of throttler checks -- we attempt to throttle once per row, which results in millions of checks per second. We need a more efficient way to keep the throttler from only running once every 250ms.

In a more advanced systems language, I believe the correct approach would be to use a **coarse clock source** for this time calculation (i.e. `CLOCK_MONOTONIC_COARSE`, https://linux.die.net/man/2/clock_gettime), but such interface is not exposed by the Go runtime. Commit https://github.com/vitessio/vitess/commit/f5af1334c812e17bdd933d17f479d1e325274510 works around this manually by creating a global ticker, shared between all the Throttlers in a process, that increments a counter once per 250ms. This way, we can delay all the throttle checks simply by ensuring that the current "tick" is newer than the "tick" when the last throttle was performed. :sparkles:

| Before | After |
|----------|--------|
| ![image](https://user-images.githubusercontent.com/42793/116109340-f9c9ad00-a6b4-11eb-93d4-5f07a55e395d.png) | _`ThrottleCheckOK` is now missing from the profile because it uses essentially no CPU time_ |

### 5. `vstreamer.(*Plan).filter`

Last, but not least, the `filter` functionality of the replication engine seems to be allocating a very significant amount of memory to perform the filtering. Why? Because it stores the resulting slice of `Value`s into a new slice of `Values`, instead of filtering it in-place. 

https://github.com/vitessio/vitess/blob/c5cd3067aeb9d952a2f45084c37634267e4f9062/go/vt/vttablet/tabletserver/vstreamer/planbuilder.go#L116-L158

If we look at the actual filtering code, we learn that the trivial optimization, filtering in-place, is not feasible, because the ordering of the resulting slice is not always linear with the input (see `result[i] = values[colExpr.ColNum]` in line 148). Our only choice is, again, to pass in a backing slice to store the results so that it can be reused between calls. Not the most elegant refactoring, but one with a significant impact in the profile:

| Before | After |
|----------|--------|
| ![image](https://user-images.githubusercontent.com/42793/116110390-eff47980-a6b5-11eb-98d5-5278259549d2.png) | ![image](https://user-images.githubusercontent.com/42793/116110443-fb47a500-a6b5-11eb-8c5a-531ff0b598a6.png) |

### Conclusion

These 5 commits bring down the total CPU usage for our `vreplication` benchmark from `17.21s` to `8.43s`, **a 49% decrease**.

![image](https://user-images.githubusercontent.com/42793/116110985-7d37ce00-a6b6-11eb-8812-608947fafcea.png)

The profile after this round of optimizations is now dominated by MySQL client reads and GRPC serialization. There is very little we can do about MySQL because we need to read from MySQL and that's not free; and there's very little we can do to GRPC without forking it, so as far as I'm concerned this is a pretty flat profile.

Again, and sadly, these optimizations have no visible impact in wall times because of the bottleneck in MySQL insertion speed, but they have a massive impact in `vttablet` throughput as a whole because we have halved the amount of CPU cycles that the replication process uses. The key takeaway here, for those following at home, is that **reducing memory allocations in a GC language doesn't just save _memory_, it mostly saves CPU time**.

## Related Issue(s)
<!-- List related issues and pull requests: -->

- 

## Checklist
- [ ] Should this PR be backported?
- [x] Tests were added or are not required
- [x] Documentation was added or is not required

## Deployment Notes
<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->

## Impacted Areas in Vitess
Components that this PR will affect:

- [ ]  Query Serving
- [x]  VReplication
- [ ]  Cluster Management
- [ ]  Build/CI
- [ ]  VTAdmin