@vmg 2021-04-27
> Without digging into the code, how many of these optimizations are shared across standard vttablet query serving?

@derekperkins: 1 and 2 are going to have an impact; number 3 could have an impact if I moved more places of the codebase to use the new `Inplace` API -- I just haven't found any of those places that are actually hot!